Worked on Tlv Headers and Tlv parsing
I created parsing for tlv headers, worked a bit more on the parsing of
the actual tlv data, as well as some of the boilerplate around the juicy
stuff.

I am however running nto a strange issue where ~80% of my tlv frames
show consistent non-random garbage values.

Often they are in the range of (but not exactly equal to):
- type: 1071847877
 len: 1055430263