I have completed chip communication and pointcloud extraction
The code which communicated with the FMCW is now working, its run on a
seperate thread and communicates over a channel.

The TLV parsing code has also got an extra function which can be called
over a thread to listen to the channel and parse any TLV data passed
onto it.

Lastly I also implemented the ability to parse the pointcloud TLV
objects (detected objects)